Add more features to misago.cache
