+++ misago/cache/__init__.py
@@ -1,3 +1,1 @@
 default_app_config = 'misago.cache.apps.MisagoCacheConfig'
-
-CACHE_NAME = "cache_versions"
